refactor(lotto) : 패키지 위치 조정
dto 패키지 구분
