+++ src/main/java/lotto/Lotto.java
@@ -2,8 +2,8 @@ package lotto;
 
 import lotto.domain.LottoBuyer;
 import lotto.domain.LottoChecker;
-import lotto.domain.LottoNumber;
-import lotto.domain.LottoWinningNumber;
+import lotto.domain.dto.LottoNumber;
+import lotto.domain.dto.LottoWinningNumber;
 import lotto.ui.InputView;
 import lotto.ui.OutputView;
