+++ src/main/java/lotto/domain/dto/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.dto;
 
 public enum LottoRank {
 	FIFTH(3, 5_000),
