+++ src/test/java/lotto/domain/LottoCheckerTest.java
@@ -1,11 +1,11 @@
 package lotto.domain;
 
+import lotto.domain.dto.LottoNumber;
+import lotto.domain.dto.LottoRank;
+import lotto.domain.dto.LottoResult;
+import lotto.domain.dto.LottoWinningNumber;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -16,19 +16,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 class LottoCheckerTest {
 
 	@Test
-	void isEqualTest() {
-		assertThat(new LottoChecker(new LottoWinningNumber(7, 1, 2, 3, 4, 5, 6), 14000))
-				.isEqualTo(new LottoChecker(new LottoWinningNumber(7, 1, 2, 3, 4, 5, 6), 14000));
-	}
-
-	@Test
 	void getWinningResultTest() {
-		LottoChecker lottoChecker = new LottoChecker(new LottoWinningNumber(7,1,2,3,4,5,6), 1000);
+		List<Integer> lottoNumbers = Stream.of(1, 2, 3, 4, 5, 6).collect(Collectors.toList());
+		LottoChecker lottoChecker = new LottoChecker(new LottoWinningNumber(lottoNumbers, 7), 1000);
 
 		assertThat(
 				lottoChecker.getWinningResult(
-						new LottoNumber(1,2,3,4,5,6),
-						new LottoNumber(1,2,3,4,5,6)))
+						new LottoNumber(lottoNumbers),
+						new LottoNumber(lottoNumbers)))
 				.isEqualTo(new LottoResult(Collections.singletonMap(LottoRank.FIRST, 2L), 1000));
 	}
 }
