+++ src/test/java/lotto/domain/dto/LottoRankTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.domain.dto;
 
+import lotto.domain.dto.LottoRank;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
