+++ src/test/java/lotto/domain/dto/LottoResultTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.dto;
 
+import lotto.domain.dto.LottoRank;
+import lotto.domain.dto.LottoResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
